requirements.txt
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.2.1
django-js-asset==1.1.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.4
whitenoise==3.3.1
